testing set1 and they all work

+++ labs/lab1/set1problem2.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 
 int main() {
-   float num;
-   float temp;
+   int num;
+   int temp;
 
    printf("Enter tempurature in F: ");
 
-   scanf("%f", &num);
+   scanf("%d", &num);
 
    //F = (C*9/5) + 32
    temp = ((num-32) * 5) / 9;
 
-   printf("The tempurature in Celsius is: %f", temp);
+   printf("The tempurature in Celsius is: %d", temp);
    
    return 0;
 }
